telegram_bot/requirements.txt
aiohttp
aiogram
aiogram_middlewares
python-dotenv